requirements.txt
Django==4.2.1
django-cleanup==7.0.0
django-filter==23.2
django-jazzmin==2.6.0
Pillow==9.4.0
reportlab==4.0.0
tzdata==2023.3
